Split Orch class into two classes: OrchBase and Orch
