+++ orchagent/Makefile.am
@@ -21,6 +21,7 @@ orchagent_SOURCES = \
 		    main.cpp \
 		    port.cpp \
 		    orchdaemon.cpp \
+		    orchbase.cpp \
 		    orch.cpp \
 		    notifications.cpp \
 		    routeorch.cpp \
@@ -47,6 +48,7 @@ orchagent_SOURCES = \
 		    neighorch.h \
 		    notifications.h \
 		    observer.h \
+		    orchbase.h \
 		    orch.h \
 		    orchdaemon.h \
 		    pfcactionhandler.h \
